slider added
