+++ src/common/styles/scss/components/_content.scss
@@ -151,6 +151,6 @@
         width: 100%;
     }
     @include md {
-        width: calc(100% - 340px);
+        width: calc(100% - 420px);
     }
 }
