--- src/common/styles/scss/components/_passion.scss
@@ -1,68 +1,0 @@
-.passion {
-    display: block;
-    
-    @include xs {
-        width: 100%;
-    }
-    @include sm {
-        width: 100%;
-    }
-    @include md {
-        width: 300px;
-        float: right;
-        padding-left: 40px;
-        margin-left: 40px;
-        border-left: 1px solid $background-color-dark;
-    }
-
-    .passion_icon {
-        height: 170px;
-        border-top: 4px solid $first-color;
-        border-left: 4px solid $first-color;
-        border-bottom: 4px solid $second-color;
-        border-right: 4px solid $second-color;
-        position: relative;
-        margin-top: 20px;
-        
-        div {
-            text-shadow: 2px 2px $highlight;
-            font-family: $display_font;
-            font-size: 24px;
-            line-height: 110%;
-            left: 0px;
-            display: inline-block;
-            position: absolute;
-            padding: 2px 4px;
-            background-color: rgba($color: $highlight, $alpha: 0.5);
-            border-bottom-right-radius: 4px;
-            border-top-right-radius: 4px;
-        }
-        div.top {
-            top: 4px;
-        }
-        div.bottom {
-            bottom: 28px;
-        }
-        span {
-            display: block;
-            position: absolute;
-            font-family: $text_font;
-            bottom: 0;
-            right: 0;
-            background-color: $background-color-dark;
-            color: $font-color-light;
-            font-size: 12px;
-            padding: 2px 5px;
-            line-height: 100%;
-        }
-    }
-    a:nth-child(1), :nth-child(4) div {
-        color: $first-color;
-    }
-    a:nth-child(2), a:nth-child(5) div {
-        color: $second-color;
-    }
-    a:nth-child(3), a:nth-child(6) div {
-        color: $third-color;
-    }
-}
