+++ src/style.scss
@@ -53,8 +53,8 @@ Components
 @import "common/styles/scss/components/homebutton";
 @import "common/styles/scss/components/instagram";
 @import "common/styles/scss/components/layout";
-@import "common/styles/scss/components/passion";
 @import "common/styles/scss/components/sidebar";
+@import "common/styles/scss/components/slideshow";
 @import "common/styles/scss/components/socialicons";
